requirements.txt
langchain-groq
crewai_tools
crewai==0.28.8 
crewai_tools==0.1.6
streamlit